requirements.txt
fastapi
pymongo
uvicorn
python-dotenv
bcrypt
pyjwt
python-jose[cryptography]  
passlib[bcrypt]  
motor
pydantic[email]
python-multipart
mangum
fastapi-cors